+++ src/main/java/pl/tw/ksiegowosc/mapper/MeritFieldSection.java
@@ -1,0 +1,8 @@
+package pl.tw.ksiegowosc.mapper;
+
+public enum MeritFieldSection {
+    CUSTOMER,
+    HEADER,
+    LINE,
+    TAX
+}